M5.0a — Rename wave_engine.py → lagrangian_engine.py + alias ewave → lagrange
The engine is no longer just a wave kernel — it will integrate a Lagrangian-derived PDE that handles wave propagation AND topology preservation via V(ψ). Module name and import alias updated to reflect M5's full scope.
